Move tests to `tests/components` directory
